Move log support for http clients to provider util package
